background snapshot: add helpers to manage a copy of ram block list
The VM ramblock list may be changed during the snapshotting.
We want to make sure that we have the same ramblock list as it
was at the time of the snapshot beginning.
For this purpose, create a copy of the ramblock list at the beginning
of the snapshotting and use it for ram saving.