blk: postpone request execution on a context protected with "drained section"
At the time, the "drained section" doesn't protect Block Driver State
from the requests appearing in the vCPU threads.
This could lead to data loss because of request coming to an unexpected BDS.

For example, when a request comes to ide controller from the guest,
the controller creates a request coroutine and executes the coroutine
in the vCPU thread. If another thread(iothread) has entered the
"drained section" on a BDS with bdrv_drained_begin, which protects
BDS AioContext from external requests, and released the AioContext
because of finishing some coroutine by the moment of the request
appearing at the ide controller, the controller acquires the AioContext
and executes its request without any respect to the entered
"drained section" producing any kinds of data inconsistency.

The patch prevents this case by putting requests from external threads to
the queue on AioContext while the context is protected for external requests
and executes those requests later on the external requests protection removing.

To distinguish external (guest) requests, those which come from the devices,
from internal ones coming from block jobs like mirroring the patch uses a new
flag to mark guest requests.

For now, we only fix ide-like devices as they issue reuests from the vCPU
context. Virtio-scsi is not needed to be fixed as they are blocked in drained
section by not processing virtio fd-s.

The patch aims to fix only IDE device in the faivor of description above by
modifing dma_blk_[read|write].
As a consecuence, it also has to modify the divces which are clients of those
dma_blk functions.

For discard and flush requests, the patch itroduces guest's versions of
flush, aio_flush, aio_pdiscard functions to ease the further fixing and
maintanance of similar cases.

How to reproduce:
1. start vm with an ide disk and a linux guest
2. in the guest run: dd if=... of=... bs=4K count=100000 oflag=direct
3. (qemu) drive_mirror "disk-name"
4. wait until block job can receive block_job_complete
5. (qemu) block_job_complete "disk-name"
6. blk_aio_p[read|write]v may appear in vCPU context (here is the race)

Signed-off-by: Denis Plotnikov <dplotnikov@virtuozzo.com>